requirements.txt
openai 
psycopg2-binary
requests
bs4
python-docx
twillio
python-dotenv